Fix adding mapped attributes
According to the previous implementation, we were able to add the mapped attributes only from the UI. But it should allow adding the mapped attributes for a claim from the claim-config.xml configuration file.

So this PR includes the fix for add the secondary user store attributes as mapped attributes from the claim-mapping.xml configuration file.

Resolves #1416