pr #1423 - Supporting cache update in addition to cache remove. Support Cluster-wide local cache invalidation support
At the moment, there is no way to invalidate the local cache in other cluster nodes. We have to wait till the cache expiry time for that.

This feature provides support for immediate invalidation of the local caches in all the in a cluster when one node removes/modifies a cache entry.