Fix governance artifacts not showing in distributed nodes when skip cache enable
This fix address the goverance artifacts not getting indexed properly when skip cache is enabled. This behavior occurs when there are
two or more nodes in deployment without clustering enabled and skip cache is set to true. The reason for this behavior is permissions
are getting cached in both the nodes. Even though skip cache is enabled, the stale permissions getting indexed when registry indexing
triggered for the resource. To overcome this issue, embedded registry get method checks specified path is added to no cache path and
refresh the permissions for given resource without fully loading the permission tree