Read tenant-cache from tenant-based CacheManager
As per the CacheManagerFactoryImpl, the CacheManager for each tenant is contained in a hash-map keyed by the tenant domain. Therefore, if the tenant is modified from within the super tenant carbon context, an obsolete CacheManager mapped to the super tenant domain will be updated, causing 2 CacheManagers per tenant.To overcome this, we are executing a temporary tenant flow to access the correct CacheManager for the tenant.

Public Issue : wso2/product-ei#1107