Fix for https://github.com/wso2/product-apim/issues/2766
Description:
Message exchange pattern property specified in the message context level is not honoured in a multi tenant scenario for JMS transport. It always gets overridden with the default IN-OUT exchange pattern. This only occurs in a multi tenant scenario. This works at super tenant level.

This can be recreated when attempting to use APIM as a JMS producer as outlined in[1]

[1] https://wso2.com/library/articles/2015/10/article-how-to-enable-wso2-api-manager-to-work-as-a-jms-producer/