Fix tenantClientService become faulty in server startup
Axis2 service tenantClientService becomes faulty in server startup,
due to this the APIs deployed in tenants cannot be invoked.
The reason for this is tenantClientService uses the SuperTenants
Axis2 Configuration so VFS listener is added to transport-ins
but the required parameters are not defined.
Fix this by removing vfs transport listner in axis2 configuration.

Fixes: https://github.com/wso2/product-ei/issues/4730