Improve the logic of generating the paginated users
With the previous implementation, we increase the pageIndex even if that page does not show users equal to pageSize. Usually, page should have users equal to the page size (Except for the last page). With this improvement, pageIndex will only increment if the page shows users equal to the page size.