Read the system property for enabling the correlation logs
Added a check to see whether the correlation logging is enabled using
the system parameter and if it is set, the configuration loaded from
the carbon.xml will be overwritten by the default enable values.

Add support to enable/disable correlation logs during runtime

The correlation log manager will now save the instances of the
log services which are responsible for handling the respective
dataholders of the correlation components.
A method is added to get the configuration settings from the
log services.

Add a correlation log service holder to manage log services

A correlation log service holder is added. It will store the instances
of the correlation log services and provide methods to access and
manipulate those log services.

Fix suggested changes in the Code Review

Log Manager reads the denied threads from the system properties if
the denied threads property and the enable correlation logs property
is available.

Fix formatting issues

Remove the default correlation component list

If the system property for enabling correlation logs is set, the
correlation log manager will not check for the component list. It
will enable all the components.

Add isDebugEnabled check before log debug