Add tenant domain validation for API requests
When creating a tenant domain, we have a validation which allows only lowercase letters, numbers, '.', '-' and '_'.
Since we are transforming the domain to lowercase when we store and query data from the database, compilations were observed when we send an API request with an uppercase tenant domain.
Therefore the same validation is added for requests as well.

Git issue: https://github.com/wso2/api-manager/issues/1466